🎨 : 회원가입2-공용 컴포넌트 사용해서 구조 변경 (예전 파일 일단 남겨둠)
